make sure MSVC doesn't complain about the implicit conversions

+++ src/file.c
@@ -14,7 +14,7 @@ u16 fgetw(FILE* fp) {
   data[0] = fgetb(fp);
   data[1] = fgetb(fp);
 
-  return (data[1] << 8) | data[0];
+  return (u16)((data[1] << 8) | data[0]);
 }
 
 /* another helper from https://github.com/viiri/st2play */
@@ -27,7 +27,7 @@ u32 fgetl(FILE *fp)
 	data[2] = fgetb(fp);
 	data[3] = fgetb(fp);
 
-	return (data[3] << 24) | (data[2] << 16) | (data[1] << 8) | data[0];
+	return (u32)((data[3] << 24) | (data[2] << 16) | (data[1] << 8) | data[0]);
 }
 
 u16 fputw(u16 value, FILE* fp) {
@@ -36,7 +36,7 @@ u16 fputw(u16 value, FILE* fp) {
   data[0] = value & 0xFF;
   data[1] = value >> 8;
 
-  return fwrite(data, sizeof(u8), 2, fp);
+  return (u16)fwrite(data, sizeof(u8), 2, fp);
 }
 
 u32 fputl(u32 value, FILE* fp) {
@@ -47,5 +47,5 @@ u32 fputl(u32 value, FILE* fp) {
   data[2] = (value >> 16) & 0xFF;
   data[3] = value >> 24;
 
-  return fwrite(data, sizeof(u8), 4, fp);
+  return (u32)fwrite(data, sizeof(u8), 4, fp);
 }
